requirements_lock.txt
aiohttp==3.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
attrs==25.4.0
boto3==1.34.0
botocore==1.34.162
certifi==2026.1.4
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.4
distlib==0.4.0
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.109.0
filelock==3.19.1
flake8==7.0.0
frozenlist==1.8.0
google-ai-generativelanguage==0.4.0
google-api-core==2.29.0
google-auth==2.48.0
google-cloud-dialogflow==2.27.0
google-generativeai==0.3.2
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.15
idna==3.11
iniconfig==2.1.0
jmespath==1.1.0
mccabe==0.7.0
multidict==6.7.1
nodeenv==1.10.0
openai==1.12.0
opensearch-py==2.4.0
packaging==26.0
platformdirs==4.4.0
pluggy==1.6.0
pre_commit==4.3.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==4.25.8
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.2.0
pytest==7.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
PyYAML==6.0.3
requests==2.31.0
rsa==4.9.1
s3transfer==0.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.35.1
tomli==2.4.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.20
uvicorn==0.27.0
virtualenv==20.36.1
yarl==1.22.0
